Update for Bind Mode
Probably the last large-scale commit for the summer... Now the user does not need to specify DNS in Bind Mode -- instead, the program will read it from the origin file. But specifying DNS feature is still available in commandline mode. The old console GUI might get deprecated one day in the future